refactor(client): extract useEditTheaterForm hook from EditTheaterModal
CRAP 172 → <50 by extracting all form logic (validation, submission,
error handling, change detection) into a dedicated hook.

EditTheaterModal is now a thin rendering shell: 320 → 207 lines.

Closes #1146
